% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkZips}
\alias{checkZips}
\title{Checks the avialability and validity of ZIP's}
\usage{
checkZips(x)
}
\arguments{
\item{x}{\code{character}, vector with full path/filename.zip.}
}
\value{
\code{logical vector} with \code{TRUE} if a zip archive exists & is not corrupt
according to \code{unzip -t} call.
}
\description{
This function uses system calls of \code{unzip -t} to check the validity of
zip archives
}
\author{
Matteo Mattiuzzi
}
