% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeQuery.R
\name{composeURL}
\alias{composeURL}
\title{Compile the URL to query the CLMS API}
\usage{
composeURL(
  productType = c("FSC", "RLIE", "PSA", "PSA-LAEA", "ARLIE"),
  geometry,
  publishedAfter,
  publishedBefore,
  startDate,
  endDate,
  productIdentifier,
  cloudCover = 100,
  queryURL,
  maxRecords = 1000
)
}
\arguments{
\item{productType}{\code{character}. One of: FSC, RLIE, PSA, PSA-LAEA, ARLIE}

\item{geometry}{\code{character}, region of interest, defined as WKT string
(POINT, POLYGON, etc.) in WGS84 projection.}

\item{publishedAfter}{\code{character}. Data published after: YYYY-MM-DD}

\item{publishedBefore}{\code{character}. Data published before: YYYY-MM-DD}

\item{startDate}{\code{character}. Sensing Date after: YYYY-MM-DD}

\item{endDate}{\code{character}. Sensing Date before: YYYY-MM-DD}

\item{productIdentifier}{\code{character}. Find products using elements found
in the filename: FSC_20170913T114531_S2B_T29UNV_V001_0: e.g. FSC_2020, TileID,
etc. (its not regex though)}

\item{queryURL}{\code{character}. Built-in HTTP query from HR-S&I portal.
If no query is provided, other the othe paramseters are used.}

\item{q}{\code{logical}. \code{TRUE}. Convert character string be converted to \code{POSIXlt}?}

\item{page}{\code{integer}. \code{TRUE}. Convert character string be converted to \code{POSIXlt}?}

\item{textualSearch}{\code{character} e.g. \code{Winter in Finland}}
}
\value{
Returns an URL \code{character} that can be submitted to the Server API.
}
\description{
This function compiles the URL statement and submits it to the API
}
\examples{

}
\author{
Matteo Mattiuzzi
}
