% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{downloader}
\alias{downloader}
\title{Downloads files from CLMS server}
\usage{
downloader(x, rootDir = "./", user, password)
}
\arguments{
\item{x}{\code{list}. The return of \code{\link{selector}} containint the
files to be donwloaded.}

\item{rootDir}{\code{character}. Default is current directory. Base location for the local archive. Inside
rootDir the function then creates a file structure similar to the one on the
server: e.g. (currently hard coded):
CLMS/Pan-European/HRL/Snow/FSC/2020/08/18/FSC_20200818T114635_S2B_T29UNV_V100_1.zip}

\item{user}{\code{character}. Your username on \url{https://cryo.land.copernicus.eu/finder/}}

\item{password}{\code{character}. Your password on \url{https://cryo.land.copernicus.eu/finder/}}
}
\value{
\code{vector} with path/filename to the locally stored zip archives.
}
\description{
This function deals with the authenification and download the files.
(currently a bit hardcoded, will be improved soon)
}
\examples{
\dontrun{
x <- composeUrl(productType = 'FSC', productIdentifier = 'T29UNV',
startDate='2020-07-01', completionDate='2020-07-31')
if(require(rjson)){
x <- fromJSON(file = x)
x <- selector(x)
downloader(x, user='yourUser', password='yourPassword')
}
}

}
\author{
Matteo Mattiuzzi
}
