% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractDate}
\alias{extractDate}
\title{Extracts the date from the filename}
\usage{
extractDate(x, split = "_", position = 2, as.POSIXlt = TRUE)
}
\arguments{
\item{x}{filename}

\item{split}{argument to isolate the date information default \code{'_'}}

\item{position}{\code{numeric} default 2. Set the position of the date after 
the \code{split} was applied}

\item{as.POSIXlt}{default \code{TRUE}}
}
\value{
If \code{as.POSIXlt == FALSE} a \code{character} string with the extracted date part
If \code{as.POSIXlt == TRUE} a sting of POSIXlt objects string.
}
\description{
Extract the date in the filename and converts it into a POSIX object.
}
\author{
Matteo Mattiuzzi
}
