% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{geo2char}
\alias{geo2char}
\title{Function to convert several extent informations into API parameter 'geometry'}
\usage{
geo2char(x)
}
\arguments{
\item{x}{one of the following classes: \code{Raster}, \code{Extent}, \code{sf},
\code{map} or \code{sp}}
}
\value{
\code{character} string in the format: \code{'POINT(lon+lat)'} or 
\code{'POLYGON(lon+lat,lon+lat,lon+lat,....)'}
}
\description{
This function will most probably become internal and is meant to convert several
R classes with a geographical extent information into the format as needed by the 
'geometry' parameter (see value).
}
\author{
Robin Lovelace and Matteo Mattiuzzi
}
