% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{listUnzip}
\alias{listUnzip}
\title{Basic wrapper around utils::unzip to allow unzippping of mutiple files.}
\usage{
listUnzip(zipfile, exdir)
}
\arguments{
\item{zipfile}{vector containing path to zipfiles}

\item{exdir}{default \code{dirname(zipdir)}}
}
\value{
\code{character} string with \code{exdir} pathnames.
}
\description{
This function allows to unzip a verctor of zip files at once.
}
\author{
Matteo Mattiuzzi
}
