% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractDate}
\alias{extractDate}
\title{Helper Functions}
\usage{
extractDate(x, unique = FALSE, as.POSIXlt = TRUE)
}
\arguments{
\item{x}{\code{character}, vector with filenames and paths returned by the server.}

\item{unique}{\code{logical}. \code{FALSE}. Should the result be collapsed using \code{\link{unique()}}}

\item{as.POSIXlt}{\code{logical}. \code{TRUE}. Convert character string be converted to \code{POSIXlt}?}
}
\value{
If \code{as.POSIXlt=TRUE} the function returns class \code{POSIXlt}, else a \code{character}.
}
\description{
This function extracts the Date from CLMS filenames and from paths returned by the API.
}
\examples{

}
\seealso{
\code{\link{as.POSIXlt}}
}
\author{
Matteo Mattiuzzi
}
