% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlhandlers.R
\name{selector}
\alias{selector}
\title{Extract information from API reply}
\usage{
selector(x, as.data.frame = FALSE)
}
\arguments{
\item{x}{\code{character}. The return of \code{\link{composeUrl}} or the REST
query generated from the portal \url{https://cryo.land.copernicus.eu/finder/}.}

\item{as.data.frame}{\code{logical}. FALSE. If true the return is converted to
a data.frame}
}
\value{
\code{list} or data.frame if \code{as.data.frame==TURE}
}
\description{
Extract file location information (sources and dests) from the API reply
}
\examples{
x <- composeUrl(productType = 'FSC', productIdentifier = 'T29UNV')
stopifnot(require(rjson))
x <- fromJSON(file = x)
selector(x)

}
\author{
Matteo Mattiuzzi
}
